tijmenw's sexy click to move is now implemented in the AI testing scene
changed the update function to a public void and then moved the call to
the "get input" scrip just so all of our "getKey" commands are in the
same place
