+++ Assets/FromLions Sucks at GameDev/Scripts/PLayer/Move_To_Click_NoNavMesh.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 4049d48250c9d424d84bc2b296e0555d
+timeCreated: 1449235979
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
