+++ ProjectSettings/TagManager.asset
@@ -22,7 +22,7 @@ TagManager:
   - Targetable Objects
   - 
   - 
-  - 
+  - scenery
   - 
   - 
   -
